fix(ci): split agent-env compaction helpers into a stdlib-only module
The compaction helpers tested by test_unit_agent_env_compaction.py
(added in PR #55) lived in agent_environment.py, which transitively
imports camel.toolkits → numpy → torch. The CI workflow installs only
the thin offline-test dep set, so collection has been failing on main
since #55 landed (ModuleNotFoundError on numpy then torch).

Move the helpers (_parse_ts / _comment_score / _compact_post_for_agent /
_compact_comment / _compact_posts_for_agent and _MAX_COMMENTS_PER_POST)
into wonderwall/social_agent/_env_compact.py — pure stdlib (json,
datetime). agent_environment.py re-exports them so its own callers are
unchanged. The test now imports from _env_compact, sidestepping the
camel chain entirely.

Net diff vs main: new module + import-rewrite in agent_environment +
test import update + workflow file unchanged.
