+++ .github/workflows/tests.yml
@@ -39,8 +39,7 @@ jobs:
             "pydantic>=2.0.0" \
             "requests>=2.31.0" \
             "neo4j>=5.15.0" \
-            "pillow>=12.0.0" \
-            "numpy>=1.26.0"
+            "pillow>=12.0.0"
 
       - name: Run unit tests
         run: pytest -m "not integration"
