+++ backend/wonderwall/social_agent/agent_environment.py
@@ -19,117 +19,16 @@ from abc import ABC, abstractmethod
 from datetime import datetime
 from string import Template
 
+from wonderwall.social_agent._env_compact import (
+    _MAX_COMMENTS_PER_POST,
+    _compact_comment,
+    _compact_post_for_agent,
+    _compact_posts_for_agent,
+    _comment_score,
+    _parse_ts,
+)
 from wonderwall.social_agent.agent_action import SocialAction
 from wonderwall.social_platform.database import get_db_path
-
-# Cap on comments per post in the agent-facing wire format. Top-K by score
-# preserves the conversation signal the agent uses for engagement decisions
-# (popular replies steer the discussion); the long tail rarely changes a
-# like / comment / repost call.
-_MAX_COMMENTS_PER_POST = 3
-
-
-def _parse_ts(ts):
-    if not ts:
-        return None
-    try:
-        return datetime.fromisoformat(str(ts).replace(' ', 'T')[:26])
-    except (ValueError, TypeError):
-        return None
-
-
-def _comment_score(c: dict) -> int:
-    return c.get('score', c.get('num_likes', 0) - c.get('num_dislikes', 0))
-
-
-def _compact_post_for_agent(p: dict, now: datetime | None) -> dict:
-    """Strip per-post fields that don't carry signal for agent decisions.
-
-    CAMEL's ChatAgent accumulates env dumps across rounds, so each post's
-    wire format is paid for many times in subsequent LLM calls. Three
-    changes here that don't change what the agent semantically sees:
-
-    - created_at → relative offset against the most recent post (e.g. "5m"),
-      since absolute timestamps in a synthetic-time sandbox carry no signal
-    - comments capped at top-K by score; total count preserved as a hint
-    - drop num_shares / num_reports / num_likes / num_dislikes when 0
-
-    Net ~30-40% fewer bytes on a typical multi-post env dump, additive with
-    the compact json.dumps in get_posts_env. Validated end-to-end on the
-    miroshark-api codebase: 57% reduction in avg input tokens per simulate
-    LLM call, 27% drop in absolute simulate cost on a 32-agent run, no
-    quality regression in the report.
-    """
-    def _delta(ts) -> str | None:
-        t = _parse_ts(ts)
-        if not t or not now:
-            return None
-        secs = max(0.0, (now - t).total_seconds())
-        if secs < 60:
-            return 'now'
-        m = int(secs // 60)
-        if m < 60:
-            return f'{m}m'
-        h = m // 60
-        return f'{h}h'
-
-    out: dict = {
-        'post_id': p.get('post_id'),
-        'user_id': p.get('user_id'),
-        'content': p.get('content'),
-    }
-    age = _delta(p.get('created_at'))
-    if age is not None:
-        out['created_at'] = age
-    elif p.get('created_at') is not None:
-        out['created_at'] = p['created_at']
-
-    if 'score' in p:
-        if p['score']:
-            out['score'] = p['score']
-    else:
-        if p.get('num_likes', 0):
-            out['num_likes'] = p['num_likes']
-        if p.get('num_dislikes', 0):
-            out['num_dislikes'] = p['num_dislikes']
-    if p.get('num_shares', 0):
-        out['num_shares'] = p['num_shares']
-    if p.get('num_reports', 0):
-        out['num_reports'] = p['num_reports']
-
-    cmts = p.get('comments') or []
-    if cmts:
-        total = len(cmts)
-        kept = sorted(cmts, key=_comment_score, reverse=True)[:_MAX_COMMENTS_PER_POST]
-        out['comments'] = [_compact_comment(c) for c in kept]
-        if total > len(kept):
-            out['comments_total'] = total
-    return out
-
-
-def _compact_comment(c: dict) -> dict:
-    out: dict = {
-        'comment_id': c.get('comment_id'),
-        'user_id': c.get('user_id'),
-        'content': c.get('content'),
-    }
-    if 'score' in c:
-        if c['score']:
-            out['score'] = c['score']
-    else:
-        if c.get('num_likes', 0):
-            out['num_likes'] = c['num_likes']
-        if c.get('num_dislikes', 0):
-            out['num_dislikes'] = c['num_dislikes']
-    return out
-
-
-def _compact_posts_for_agent(posts: list) -> list:
-    if not posts:
-        return posts
-    valid_ts = [t for t in (_parse_ts(p.get('created_at')) for p in posts) if t]
-    now = max(valid_ts) if valid_ts else None
-    return [_compact_post_for_agent(p, now) for p in posts]
 
 
 class Environment(ABC):
